requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
marshmallow==3.20.1
python-dotenv==1.0.0
pytest==7.4.2
black==23.9.1
isort==5.12.0
flake8==6.1.0